Move scripts
