+++ test_scripts/Security/GetSystemTime/common.lua
@@ -7,7 +7,7 @@ local security = require("user_modules/sequences/security")
 local utils = require("user_modules/utils")
 local test = require("user_modules/dummy_connecttest")
 local commonFunctions = require("user_modules/shared_testcases/commonFunctions")
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ General configuration parameters ]]
 config.serverCertificatePath = "./files/Security/GetSystemTime_certificates/spt_credential.pem"
