+++ test_scripts/Security/SSLHandshakeFlow/002_Navi_Predefined_cert_expired_PTU_cert_valid_SUCCESS.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
@@ -14,7 +14,7 @@ config.application1.registerAppInterfaceParams.appHMIType = { "NAVIGATION" }
 --[[ Local Variables ]]
 
 --[[ Local Functions ]]
-local function startServiceProtectedACK()
+local function startServiceSecuredACK()
   local serviceId = 7
   common.getMobileSession():StartSecureService(serviceId)
   common.getMobileSession():ExpectControlMessage(serviceId, {
@@ -23,18 +23,24 @@ local function startServiceProtectedACK()
   })
   common.getMobileSession():ExpectHandshakeMessage()
   :Times(1)
+  local function ptUpdate(pTbl)
+    local filePath = "./files/Security/client_credential.pem"
+    local crt = common.readFile(filePath)
+    pTbl.policy_table.module_config.certificate = crt
+  end
+  common.policyTableUpdateSuccess(ptUpdate)
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Clean-up certificates", common.cleanUpCertificates)
-runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential.pem" })
+runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
 runner.Step("Register App", common.registerApp)
-runner.Step("Switch RPC Service to Protected mode ACK", startServiceProtectedACK)
+runner.Step("PolicyTableUpdate", common.policyTableUpdate)
+runner.Step("Switch RPC Service to Protected mode ACK", startServiceSecuredACK)
 runner.Step("Activate App Protected", common.activateAppProtected)
 runner.Step("AddCommand Protected", common.sendAddCommandProtected)
