+++ test_scripts/Security/SSLHandshakeFlow/004_Navi_Predefined_cert_expired_PTU_cert_expired_NACK.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
@@ -14,17 +14,17 @@ config.application1.registerAppInterfaceParams.appHMIType = { "NAVIGATION" }
 --[[ Local Variables ]]
 
 --[[ Local Functions ]]
-local function startServiceSecuredACK()
+local function startServiceSecuredNACK()
   local serviceId = 7
   common.getMobileSession():StartSecureService(serviceId)
   common.getMobileSession():ExpectControlMessage(serviceId, {
-    frameInfo = common.frameInfo.START_SERVICE_ACK,
-    encryption = true
+    frameInfo = common.frameInfo.START_SERVICE_NACK,
+    encryption = false
   })
   common.getMobileSession():ExpectHandshakeMessage()
-  :Times(1)
+  :Times(0)
   local function ptUpdate(pTbl)
-    local filePath = "./files/Security/client_credential.pem"
+    local filePath = "./files/Security/client_credential_expired.pem"
     local crt = common.readFile(filePath)
     pTbl.policy_table.module_config.certificate = crt
   end
@@ -34,15 +34,13 @@ end
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem", false })
+runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
 runner.Step("Register App", common.registerApp)
 runner.Step("PolicyTableUpdate", common.policyTableUpdate)
-runner.Step("Switch RPC Service to Protected mode ACK", startServiceSecuredACK)
-runner.Step("Activate App Protected", common.activateAppProtected)
-runner.Step("AddCommand Protected", common.sendAddCommandProtected)
+runner.Step("Switch RPC Service to Protected mode NACK", startServiceSecuredNACK)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL, clean-up certificates", common.postconditions)
