+++ test_scripts/Security/SSLHandshakeFlow/005_Navi_Predefined_cert_missing_PTU_cert_valid_SUCCESS.lua
@@ -5,26 +5,28 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
-config.application1.registerAppInterfaceParams.appHMIType = { "DEFAULT" }
+config.application1.registerAppInterfaceParams.appHMIType = { "NAVIGATION" }
 
 --[[ Local Variables ]]
 
 --[[ Local Functions ]]
-local function startServiceSecuredNACK()
+local function startServiceSecuredACK()
   local serviceId = 7
   common.getMobileSession():StartSecureService(serviceId)
   common.getMobileSession():ExpectControlMessage(serviceId, {
-    frameInfo = common.frameInfo.START_SERVICE_NACK,
-    encryption = false
+    frameInfo = common.frameInfo.START_SERVICE_ACK,
+    encryption = true
   })
   common.getMobileSession():ExpectHandshakeMessage()
-  :Times(0)
+  :Times(1)
   local function ptUpdate(pTbl)
-    pTbl.policy_table.module_config.certificate = nil
+    local filePath = "./files/Security/client_credential.pem"
+    local crt = common.readFile(filePath)
+    pTbl.policy_table.module_config.certificate = crt
   end
   common.policyTableUpdateSuccess(ptUpdate)
 end
@@ -32,13 +34,15 @@ end
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem" })
+runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem", false })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
 runner.Step("Register App", common.registerApp)
 runner.Step("PolicyTableUpdate", common.policyTableUpdate)
-runner.Step("Switch RPC Service to Protected mode NACK", startServiceSecuredNACK)
+runner.Step("Switch RPC Service to Protected mode ACK", startServiceSecuredACK)
+runner.Step("Activate App Protected", common.activateAppProtected)
+runner.Step("AddCommand Protected", common.sendAddCommandProtected)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL, clean-up certificates", common.postconditions)
