+++ test_scripts/Security/SSLHandshakeFlow/008_Non-Navi_Predefined_cert_valid_SUCCESS.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
@@ -14,7 +14,7 @@ config.application1.registerAppInterfaceParams.appHMIType = { "DEFAULT" }
 --[[ Local Variables ]]
 
 --[[ Local Functions ]]
-local function startServiceSecuredACK()
+local function startServiceProtectedACK()
   local serviceId = 7
   common.getMobileSession():StartSecureService(serviceId)
   common.getMobileSession():ExpectControlMessage(serviceId, {
@@ -23,24 +23,17 @@ local function startServiceSecuredACK()
   })
   common.getMobileSession():ExpectHandshakeMessage()
   :Times(1)
-  local function ptUpdate(pTbl)
-    local filePath = "./files/Security/client_credential.pem"
-    local crt = common.readFile(filePath)
-    pTbl.policy_table.module_config.certificate = crt
-  end
-  common.policyTableUpdateSuccess(ptUpdate)
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential_expired.pem", false })
+runner.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential.pem" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
 runner.Step("Register App", common.registerApp)
-runner.Step("PolicyTableUpdate", common.policyTableUpdate)
-runner.Step("Switch RPC Service to Protected mode ACK", startServiceSecuredACK)
+runner.Step("Switch RPC Service to Protected mode ACK", startServiceProtectedACK)
 runner.Step("Activate App Protected", common.activateAppProtected)
 runner.Step("AddCommand Protected", common.sendAddCommandProtected)
