+++ test_scripts/Security/SSLHandshakeFlow/009_Non-Navi_Predefined_cert_expired_PTU_cert_valid_SUCCESS.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
