+++ test_scripts/Security/SSLHandshakeFlow/014_Non-Navi_Predefined_cert_missing_PTU_cert_expired_NACK.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require("test_scripts/Security/common")
+local common = require("test_scripts/Security/SSLHandshakeFlow/common")
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
